Evaluate a running process
