+++ client/system/info_freebsd.go
@@ -41,11 +41,6 @@ func GetInfo(ctx context.Context) *Info {
 	return gio
 }
 
-// checkFileAndProcess checks if the file path exists and if a process is running at that path.
-func checkFileAndProcess(paths []string) ([]File, error) {
-	return []File{}, nil
-}
-
 func _getInfo() string {
 	cmd := exec.Command("uname", "-sri")
 	cmd.Stdin = strings.NewReader("some input")
