+++ client/system/info_windows.go
@@ -85,11 +85,6 @@ func GetInfo(ctx context.Context) *Info {
 	return gio
 }
 
-// checkFileAndProcess checks if the file path exists and if a process is running at that path.
-func checkFileAndProcess(paths []string) ([]File, error) {
-	return []File{}, nil
-}
-
 func getOSNameAndVersion() (string, string) {
 	var dst []Win32_OperatingSystem
 	query := wmi.CreateQuery(&dst, "")
